+++ flutter_app/lib/domains/avatar/Avatar.dart
