+++ flutter_app/lib/domains/messages/SocketMessage.dart
@@ -18,10 +18,10 @@ class SocketMessage {
     int indexOfOpenBracket = message.indexOf('[');
     int indexOfOpenBrace = message.indexOf('{');
 
-    SerializedMessageType serializedMessageType = _getSerializedMessageType(message);
+    SerializedMessageType serializedMessageType =
+        _getSerializedMessageType(message);
 
     switch (serializedMessageType) {
-
       case SerializedMessageType.EMPTY:
         content = new EmptyMessageContent();
         break;
@@ -35,14 +35,16 @@ class SocketMessage {
 
         int propertyBagListLength = messageArray.propertyBagList.length;
         if (propertyBagListLength == 1) {
-          content = new SingleValueDynamicMessageContent(messageArray.propertyBagList[0]);
+          content = new SingleValueDynamicMessageContent(
+              messageArray.propertyBagList[0]);
         } else if (propertyBagListLength > 1) {
           type = messageArray.propertyBagList[0];
           if (!(content is Map)) {
-            content = new SingleValueDynamicMessageContent(messageArray.propertyBagList[1]);
-          }
-          else {
-            content = new DynamicMessageContent(messageArray.propertyBagList[1]);
+            content = new SingleValueDynamicMessageContent(
+                messageArray.propertyBagList[1]);
+          } else {
+            content =
+                new DynamicMessageContent(messageArray.propertyBagList[1]);
           }
         } else {
           content = new EmptyMessageContent();
