+++ flutter_app/lib/domains/messages/incoming-messages/ArrayMessage.dart
@@ -1,6 +1,6 @@
 import 'package:flutterapp/domains/messages/incoming-messages/ReceivedMessage.dart';
 
-class ArrayMessage extends ReceivedMessage{
+class ArrayMessage extends ReceivedMessage {
   String message;
 
   ArrayMessage(this.message) {
