+++ flutter_app/lib/domains/messages/incoming-messages/ObjectMessage.dart
@@ -6,6 +6,4 @@ class ObjectMessage extends ReceivedMessage {
   ObjectMessage(this.message) {
     this.messageType = "Object";
   }
-
-
 }
