+++ flutter_app/lib/domains/messages/incoming-messages/ServerTimeMessage.dart
@@ -1,10 +1,9 @@
 import 'package:flutterapp/domains/messages/incoming-messages/ReceivedMessage.dart';
 
-class ServerTimeMessage extends ReceivedMessage{
+class ServerTimeMessage extends ReceivedMessage {
   int serverTime = 0;
 
   ServerTimeMessage(this.serverTime) {
     this.messageType = "ServerTime";
   }
-
 }
