+++ flutter_app/lib/domains/messages/incoming-messages/UpdateMessage.dart
@@ -14,7 +14,5 @@ class UpdateMessage extends ReceivedMessage {
     this.lastKnownTimeUpdatedAt = objectFromMessage['lastKnownTimeUpdatedAt'];
     this.state = objectFromMessage['state'];
     this.videoDuration = objectFromMessage['videoDuration'];
-
   }
-
 }
