+++ flutter_app/lib/domains/messages/incoming-messages/UserIdMessage.dart
@@ -6,6 +6,4 @@ class UserIdMessage extends ReceivedMessage {
   UserIdMessage(this.userId) {
     this.messageType = "UserId";
   }
-
-
 }
