+++ flutter_app/lib/domains/messages/outgoing-messages/join-session/JoinSessionMessage.dart
