+++ flutter_app/lib/theming/AppTheme.dart
@@ -5,18 +5,19 @@ class PartyHarderTheme {
     return new ThemeData(
         primaryColor: Colors.redAccent,
         unselectedWidgetColor: Colors.black,
-        cardColor: Colors.black12
-    );
+        cardColor: Colors.black12);
   }
+
   static getDarkTheme() {
     return new ThemeData(
-        brightness: Brightness.dark,
-        primaryColor: Colors.redAccent,
-        accentColor: Colors.redAccent,
-        buttonColor: Colors.redAccent,
-        dialogBackgroundColor: Colors.blueGrey,
-        cardColor: Colors.white12,
-        unselectedWidgetColor: Colors.white60
+      brightness: Brightness.dark,
+      primaryColor: Colors.redAccent,
+      accentColor: Colors.redAccent,
+      buttonColor: Colors.redAccent,
+      dialogBackgroundColor: Colors.blueGrey,
+      cardColor: Colors.white12,
+      unselectedWidgetColor: Colors.white60,
+      canvasColor: Colors.black26,
     );
   }
 }
