+++ flutter_app/test/domains/messenger/Messenger_test.dart
@@ -6,7 +6,9 @@ import 'package:web_socket_channel/io.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
 class MockIOWebSocketChannel extends Mock implements IOWebSocketChannel {}
+
 class MockWebSocketSink extends Mock implements WebSocketSink {}
+
 class MockSocketMessage extends Mock implements SocketMessage {}
 
 void main() {
