+++ flutter_app/test/test_test.dart
@@ -1,7 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  test('that it should pass', () => {
-    expect(true, isTrue)
-  });
+  test('that it should pass', () => {expect(true, isTrue)});
 }
